Rename "wgd" module to "wgd_ksrates"
