+++ wgd_ksrates/diamond.py
